Compilation fix
git-svn-id: file:///home/roberto/Git/svn@6523 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: bf1f5e1cfa9e6ebafde827edb5b28bb4889a0050

+++ trunk/xmipp/libraries/data/basis.cpp
@@ -193,9 +193,9 @@ void Basis::changeToVoxels(GridVolume &vol_basis, MultidimArray<double> *vol_vox
             STARTINGZ(*vol_voxels) - zdiff,
             STARTINGY(*vol_voxels) - ydiff,
             STARTINGX(*vol_voxels) - xdiff,
-            FINISHINGZ(*vol_voxels) + Zdim - ZSIZE(*vol_voxels) - zdiff,
-            FINISHINGY(*vol_voxels) + Ydim - YSIZE(*vol_voxels) - ydiff,
-            FINISHINGX(*vol_voxels) + Xdim - XSIZE(*vol_voxels) - xdiff);
+            (int)(FINISHINGZ(*vol_voxels) + Zdim - ZSIZE(*vol_voxels) - zdiff),
+            (int)(FINISHINGY(*vol_voxels) + Ydim - YSIZE(*vol_voxels) - ydiff),
+            (int)(FINISHINGX(*vol_voxels) + Xdim - XSIZE(*vol_voxels) - xdiff));
         break;
     case splines:
         spatial_Bspline032voxels(vol_basis, vol_voxels, Zdim, Ydim, Xdim);
